Code cleanup, write path rewritten.
Removed all the proc functions, move_data* family
of functions.
Our checkpoint now holds the reverse translation
table: PBA, LBA. In this, we already know the PBA
from the write frontier. We need to write fewer
blocks this way when the writes are spread over
the disk as they are written sequentially by
their PBA (which is also sequentially increasing)

Signed-off-by: Surbhi Palande <csurbhi@gmail.com>