Interim work.
Changed the metadata.
Now maintaining a reverse map and then 1 sector of checkpoint
metadata to get the disk state faster. The revmap is as follows:
1st sector holds the metadata for the revmap. In that it tells
us whats the magic number for current revmap entries: 0 or 1.
We keep flipping it on every overwrite.
The next sector holds 6 entries and then is followed by a 32 bit
crc and the bit. In reality we could simply store the 1/0 and
we still know if this sector was recently written out or not.
Thus every sector holds 6 reverse map entries. This way
we can find out if this sector was written out before
a crash or is this stale data.

We also use "sector atomicity" and write pointer advancement
in data zones for verifying translations. If there is a translation
entry, the data zone pointer in that zone should be ahead
of the data for which the translation exists.

This change in metadata is now halfway encorporated in format.c
but the metadata_format.h has all the structures as needed.

When we flush data: we use the following algorithm:
1) flush the data first
2) write out the translation entries - as mentioned above.
3) when a new zone is requested - call checkpoint:

On checkpoint:
a) Write out the reverse map entries to the translation table
b) barrier
c) write out the sement entries to the SIT area
d) barrier
e) write out the checkpoint metadata.
f) barrier

Now continue with the write operation. We want to time the above
checkpoint procedure and then think of optimizations.
For now, this should work.

Signed-off-by: Surbhi Palande <csurbhi@gmail.com>