Write error handling and rev map entries flush.
Written the bones of:
a) write I/O completion.
Realized that we need to keep the status of a zone
on disk. If a write to a SMR zone fails, then consequent
writes will fail as well because the write pointer will
not have progressed in the failure write. As a result,
such a SMR zone needs to be flagged erroneous. We have
no provision for this right now. Need to think of this.
Right now, cloned a bio so that it shares the same data pages
but can have different private field, io_endio and other
metadata fields. Set the endio to point to our endio that
handles the errors and removes the translation entries from
the RB tree and from the disk writable entries. There will
be no further entries for this SMR zone. For a random zone,
the erroneous write creates a hole in the zone, but further
writes can progress. This could be a result of a bad sector.
Thus a bad sector causes a zone failure on SMR as far as writes
are concerned.

b) reverse map entries flush. We want to flush them only when
the corresponding bio write succeeds.

Signed-off-by: Surbhi Palande <csurbhi@gmail.com>