Work in progress.
Completed coding for:
a) Flushing the reverse map pages to the disk
b) flushing the entries from reverse map to the correct
place on the translation map. This is currently done
when 100 translation blocks accumulate. This number can be changed
c) Before reusing the reverse map blocks (which are a cache for the
translation blocks), we need to check if they are flushed in the
translation map blocks appropriately. Designed a bitmap to indicate
this. When all the entries of a reverse map page are flushed to disk
the appropriate bit is cleared. Such a cleared bit indicates that
reusing that block is possible. Designed the code to read, flush this
bitmap to disk

Saving interim work

Signed-off-by: Surbhi Palande <csurbhi@gmail.com>