Write simplified and corrected.
Simplified the write.
Corrected the part where when padding was required a new page was being
assigned. That is wrong, because you will never need a block worth of
pad. In fact, you do have a page and have a request to write only a
portion of the page. Since we write blocks at a time, we zero out the
remaining page and write it.
I think this zeroing out should not be necessary as the filesystem
should be taking care of this. The other callers are direct block io
callers and they too should go through vfs. The vfs write will always
give you a zeroed out page. But, for now doing this extra work.
TODO: verify this and remove the zeroing out if not necessary

Signed-off-by: Surbhi Palande <csurbhi@gmail.com>