Work in progress.
Eye balled the code and then corrected obvious errors. This is still not
complete. Yet to compile test. But design wise goes as follows:

a) map_write_io -> nstl_clone_endio() -> revmap_entries_write()
When a block worth of revmap entries are created, they are flushed
to the disk. The endio of this write creates translation map entries;
when 100 pages accumulate, they are written out to the disk. Every
translation map page can have entries for 'n' reverse map blocks. Each
reverse map block has a closure. This closure is incremented for every
translation map page that it refers to. Thus, the 'n' reverse map blocks
that create the entries in a translation map page, corresponds to n
closures. These are store in a linked list and incremented every time an
entry is added for the first time from the reverse map page (i.e closure
is unique for all the entries in ONE reverse map page). When the
translation map is written to the disk, all such collected closures are
decremented. When the value becomes zero, the reverse map page can be
freed and the reverse map block is marked for reuse.

The in memory translation table is immediately updated, as otherwise we
would not be able to read a block that had returned a successful write.
Here we risk loosing data when the reverse map entries are not flushed
and we crash! We wait for max 10 seconds, as every 10 seconds or when
1 block worth entries (24) accumulate we always write them out to the
disk.

Signed-off-by: Surbhi Palande <csurbhi@gmail.com>