formatting code added.
Current problems:
a) Translation table: 100 blocks should be shadowed before writing
them at their currently location. Absence of this, could potentially
lead to loss of data. The code framework to do this is present; it
should be similar to the code that flushes reverse map to the
translation map.
b) More importantly: read/write path has bugs, that need debugging.

Signed-off-by: Surbhi Palande <csurbhi@gmail.com>