Testing: fix for bug: scheduling idle.
The original code was as follows:

      Th1(flush revmap entries  |             Th2 (revmap_entriesflushed)
--------------------------------------------------------------------------
t1                              |  clear_bit

t2 mark_bit

t3 wait_for_bit to
    be cleared.

Deadlock.

The ctx->rev_flush_lock was not used at certain places consistently
and at one place in the following order A, A
That leads to a deadlock too.
Fixed all of these issues in this untested patch.

Signed-off-by: Surbhi Palande <csurbhi@gmail.com>