Fixed bugs related to get_new_zone.
format.c:: get_zone0_pba() get the sector address of the next block
after SIT and not the next zone like we expect.
sb->zone0_pba is the pba of the first zone available to user.
This should thus also be subtracted from get_zone_nr() so that
all user zones are numbered with this zone as the base.
Untested code.

Signed-off-by: Surbhi Palande <csurbhi@gmail.com>