TODO: merge nodes on sequential writes.
Current STL code is not extent based in the true sense.
sequential writes create different nodes.
It does not merge nodes when sequential writes are present.
This patch partially fixes that. It does not address the problem of
modifying the right nodes when the extension/merging might overlap with
the right node. The new merged node could overlap with the LBA on the
right node[s]. Keeping this patch for documentation and marking this
task as one of the many TODOS.

Signed-off-by: Surbhi Palande <csurbhi@gmail.com>