Rewrote the RB tree updates.
Rewrote the original RB tree implemetation to fix the following bugs:
a) On sequentially contiguous separate writes, nodes were not getting
merged.
The case where new node overlaps multiple existing nodes was not handled
properly for the case where the last segment is not completely
overlapped, but partly overlapped so that the last part needs to be
split such that the first half of the existing segment is overlapped
and the second half is kept as is since its PBA is different.
Tested this code by using a RB tree userspace implementation for the
kernel RB tree equivalent.

Signed-off-by: Surbhi Palande <csurbhi@gmail.com>