Adding functionality for Background GC
Added the following functionality:
a) Maintaining a RB tree for segment/zone information. This tree is
sorted on "cb_cost" (cost benefit cost). The key cost may not be unique. Non
unique nodes are identified using "zonenr" and always to the left.
This balanced tree is ready for being used in the Background GC case.
For foreground GC, we use the greedy approach. When foreground GC
triggers, we will create another tree that is sorted on the number of
valid blocks. We can use only the left half of the tree if there are
enough segments to clean. This way we apply greedy algorithm on the
older zones and try to maintain some data blocks stability. We can
compare this algorithm against the pure greedy approach. Will add the
Foreground GC code later.

For background GC:
a) we pick the left most node as it has the lowest cost.
b) we get the PBA for this node and then lookup the reverse tree to find
out the corresponding LBA, len pair. If none is found this block is
valid and we move to PBA = PBA + 1. If a pair is found, we move to
PBA = PBA + len. We continue this till all the PBAs in the zone are
exhausted.
c) TODO: Read all these valid PBAs.
d) TODO: write them out.
e) TODO: Before pointing metadata to these new PBAs, we check if any writes
updated the in memory translation tables. If they did then those PBAs
are invalid blocks and we dont update the metadata corresponding to
those PBAs. If they didn't then we do. This part is racy and needs a
valid lock that is used both by application writes and GC writes. This
part is also not a part of this patch. We follow it up in the next
patches.

Signed-off-by: Surbhi Palande <csurbhi@gmail.com>