Rearchitecting GC model
Current model:
GC thread executes and checks if system is io idle
and if so then performs gc and goes away.
In this, you can loose important instances when Background gc could be
done.

Ideally:
a) When we are ioidle, we invoke GC.
b) during GC if we see that app read/write has arrived, we stop our
work.
After the app I/O is done, we can resume the GC where we started.
So our GC loop is basically continuous. We stop when app I/O comes
and we keep working through. When memory is under pressure, we can
delete the GC data!
Discuss this model!

Commented the timer that will trigger; if it triggers then we have
confirmation that system is i/o idle. At this point we need to restart
from where we were at in the past.

After a low water mark is invoked, GC mode will be FG GC mode. In this
mode, we complete GC rather than stopping when we see an I/O. We should
try I/O priorities for the FG GC mode.

Signed-off-by: Surbhi Palande <csurbhi@gmail.com>