bio_endio cannot sleep.
Getting "scheduling while atomic bug"
Changed the bio_endio to not sleep.
The bio completion is done using the completion variables in a thread.
This thread then continues to add translation entries corresponding to
the revmap entries in the revmap page. The revmap page is deleted only
once all the corresponding translation pages that have those revmap
entries are flushed to the disk.

Signed-off-by: Surbhi Palande <csurbhi@gmail.com>