bio_endio should not sleep on write completion.
nstl_clone_endio was calling add_revmap_entries.
This function calls page_alloc and also calls a function to flush revmap
entries when a page full of them are written. This flushing could sleep
as well. So we create a thread where we wait for the IO to complete and
call these functions from this thread.

Signed-off-by: Surbhi Palande <csurbhi@gmail.com>