Device mapper version 1 works.
This patch has fixes for reference count management of revmap pages and
associated translation map:
a) there was an extra kref_get
b) the addition of nodes to linked list had an error (different strucks
were getting threaded on the linked list)
c) Do not add the nodes in a SIT based in memory tree using the CB cost
as the min mtime is not yet known. Also min_mtime changes when the
segment with the min_mtime is removed. Currently we use the Greedy cost
(ie. the number of valid blocks in a segment).
c) some other minor fixes.

Signed-off-by: Surbhi Palande <csurbhi@gmail.com>