Add a node with lba and pba smaller than any node
The in memory translation map is updated asynchronously.
This means that the nodes added need not have LBAs that are
always sequentially increasing. Merge nodes when the LBAs and PBAs align
on the left of an existing node. Without this patch, this case is not
handled and stl_update_range gets stuck in the while loop.

Signed-off-by: Surbhi Palande <csurbhi@gmail.com>