Delete or split overlapping nodes after merge.
When an existing extent is merged with a new write; we need to take more
action if this write was an overwrite i.e not the first write to this
LBA. The extents overlapping with this merged LBA, PBA, len pair should
be split/deleted accordingly. Else you will have two entries for the
same LBA leading to a corrupt tree.

TODO:
One merge could make it possible to merge exactly one more node to the
left of right.

Signed-off-by: Surbhi Palande <csurbhi@gmail.com>