Bug fix in lsdm_update_range code.
When you merge a node you can have the following case:

	-----------**********

LBA of - and + matches on the left and right end respectively.
But the PBA does not match.
Now suppose we get an overwrite on *** such that the PBA now matches,
then:
		   ++++++++++
	-----------**********

Now first the merge code will be executed so that ---- and +++ will be
merged, but ***** needs to be removed. The earlier patch had removed
this case, reinstated it.

Signed-off-by: Surbhi Palande <csurbhi@gmail.com>