Fixed some bugs in lsdm_update_range()
a) split_delete_nodes() was working from the root of the tree all over
again. While doing this, it was also splitting the newly merged nodes
and undoing its effect. We have to work with the partial trees.
when we merge from the right, we want to work with theright tree.
when we merge from the left we want to work with the left tree.

b) we were working with rb_left and rb_right initially and then working
with in sequence values. That does not work well as you could land in an
in between value and then next and prev will not cut all the cases.
So using the sequential values.

Signed-off-by: Surbhi Palande <csurbhi@gmail.com>