Fixed bugs in rb tree insert and update code.
lsdm_rb_insert() should follow these rules:
a) When we are trying to insert new, new should not overlap with any
existing node. We should have processed the overlapping nodes so that
this rule applies by now.
b) We go left only when the entire segment is smaller than the node in
consideration.
c) We go right only when the entire segments is greater than the node in
consideration.

lsdm_update_range() should follow these rules:
a) We go left only when the entire segment is smaller than the node in
consideration.
b) We go right only when the entire segments is greater than the node in
consideration.
c) When we find an overlapping node, this node is not guaranteed to be
the smallest overlapping node. This simplifies the code a lot.
So we find the smallest overlapping node and then take care of the
following cases:

Case 1: overwrite a part of the existing extent
         *      ++++++++++
         * -----------------------

Case 2: Overwrites an existing extent partially;
         * covers only the right portion of an existing extent (e)
         *      ++++++++
         * -----------
         *
         *      ++++++++
         * -------------

Case 3: Overwrite many extents completely
         *      ++++++++++++++++++++
         *        ----- ------  --------
         *
         * Could also be exact same:
         *      +++++
         *      -----

Case 4: Partially overwrite an extent
         * ++++++++++
         *      -------------- OR
         *
         * Left end of + and - matches!
         * +++++++
         * --------------

Because we started from the extent with the smallest overlapping LBA, we
do not have to loop through these cases as these will be the only cases
that will apply in increasing order.

Also, we check if nodes can be merged after we have inserted a node.
The merge code checks if the node newly added can be merged with its
previous and/or next neighbor.

Tested with 2 repeated invocations of mkfs in the userspace counterpart.

Signed-off-by: Surbhi Palande <csurbhi@gmail.com>