Use different rb tree for GC.
We cannot piggy back on the sit info as that is sorted on blocknr that
holds the segment information (One block holds info of multiple
segments). We create a new RB tree for GC information; one node holds
information of ONE segment. The states of a segment are:
a) EMPTY: nrblks == 65536
b) FULL: nrblks = 0, and the current segment is not this segment.
c) NEEDS GC: nrblks < 65536 and > 0, and the current segment is not this
segment.
Our GC tree holds a node, that maintains cost of doing this GC, nrblks.
When the cost is the same for two nodes, then the nrblk must be
different as mtime is always different. When mtime is small, nrblks must
be big. We prefer old segments when the cost for cleaning is the same.

Signed-off-by: Surbhi Palande <csurbhi@gmail.com>