Fixed some bugs.
a) lsdm_search_rb* had a copy by value error, where **ptr was passed
but ptr was being set. This change wont get reflected in the caller.
Changed this function.

b) Changed lsdm_update_range() to reflect the  change in a.

c) Changed the wait_on_completion() to wait_on_completion_timeout().
Some I/Os just take too long and we cannot wait for that much time.
We can see these i/o reflected as "Buffer I/O error * " messages.

d) For now removed all the waiting functions. When we send i/o to the
block, we rely on that.

Signed-off-by: Surbhi Palande <csurbhi@gmail.com>