Debugging gc process.
reverse map table stores the extent in the opposite way of map.
Thus an extent->lba is actually the pba in the reverse map.
This decision was made so that the code could be reused.
In the future we will rever this decision as code readability is more
important than reusability.

Signed-off-by: Surbhi Palande <csurbhi@gmail.com>