Debugging: adding same lba twice.
process_tm_entries() is meant to add a page with existing translation
entries to memory. This is updated from revmap entries. If no
translation map page exists then such a page is added.

read_extents_from_block() is called at boot time to create the extent
tree. From then on the tree is updated on every write. We do not need to
call it here, as the extent tree already has all these LBAs.

Signed-off-by: Surbhi Palande <csurbhi@gmail.com>