Translation maps flushed after revmap.
process_tm() gets called before flush_translation_blocks() as we add a
flush_workqueue() call before flush_translation_blocks().
We also moved the wakeup after the work is queued in the bio_endio of
revmap_flush call. The wait in the dtr thus gets called only after the
work to process_tm is queued. This is followed by flush_workqueue() that
waits for the completion of every work function that was queued!!

Signed-off-by: Surbhi Palande <csurbhi@gmail.com>