Fixed the 75zones writes only problem.
dmsetup create was getting configured with 75 zones.
The command and its parameters were printed at the  end
of the format program. This was initially done for testing.
I forgot all about it later and kept using what was
outputed by format program. Changed the default 75 to use
the number of zones on the ZNS device.

Signed-off-by: Surbhi Palande <csurbhi@gmail.com>