Synchronous calling of flushing metadata.
The metadata flushing is getting called from an asynchronous queued work
that gets called when write is completed successfully. This flush
operation does not have to be asynchronous as we are not in the context
of write and are not in the application's path! Calling the flush
function makes it easier to debug and follow the code.

Signed-off-by: Surbhi Palande <csurbhi@gmail.com>