Fixed reverse map - duplicate extent addition.
extents are added in the translation map based on "lba" as the key. The
reverse map on the other hand uses "pba" as the key. While the addition
can be similar, the logic for next node and previous node will always be
based on lba. Thus we cannot reuse the code. As a side effect, made the
addition of reverse map extents easier; maintaining a pointer in both
the extents that points to each other. When an extent in a translation
map is removed, the corresponding extent in the reverse map is removed.
When it is added, an extent is added in the reverse map. Nothing needs
to be done when an extent in the translation map is adjusted as the
extent in the reverse map only maintains the pba and not the lba and the
length. This is because, the lba and len can easily be found from the
translation map! This codes ensures that the reverse translation map and
the translation map are in sync with each other!

Signed-off-by: Surbhi Palande <csurbhi@gmail.com>