Starting the flusher thread.
Earlier the flusher thread was using the queue where the gc thread was
parked. When the flusher thread timer was triggered, it woke up whatever
process was at the head of the queue. Thus gc was getting triggered from
the flusher thread.
Using a separate flusher queue.

Also, had forgotten to mark the revmap bit map bit - used and free when
the revmap block is flushed and when the translation entries in the
revmap map are flushed to the translation table respectively.

Signed-off-by: Surbhi Palande <csurbhi@gmail.com>