Cleaned up the GC RB tree code.
Linux kernel RB tree gets corrupted if the key for the RB tree is not
unique. So, we maintain two separate trees: a zone tree and a cost tree.
When multiple zones have the same cost, we maintain a linked list in the
cost tree node that maintains all these zones. The zone tree node
maintains a ptr to the cost tree node for tracking. So we add a zone to
the zone tree node and when its time to remove, we remove the node from
the zone tree and from the linked list maintained in the cost node. If
this is the only zone on the linked list, then we delete the linked in
the cost node and then also delete the cost node.

Tested code by writing one zone of blocks and overwriting alternately.

Signed-off-by: Surbhi Palande <csurbhi@gmail.com>