Test with a zone count of 4032.
For our research purpose, we do the testing with a disk size of 1TB.
The reason for this is that, we are currently writing the TM blocks in
the CMR zones. There are only 64 non shingled zones available, we use 63
for TM data that is page based. We have a translation entry for every
4096 blocks. We store only the PBA in the entry as the LBA is calculated
from the offset. Thus only 64 entries are stored in one block. 63 zones
have 63 * 65536 blocks, thus total (63 * 65536 * 64) entries can be
stored in the non shingled zones. That leads to translation entries
worth (63 * 64) zones ie 4032 zones. 4032 zones is nearly 1TB of data.
Its actually 64 zones short, as 4096 zones is 1TB of data.

Signed-off-by: Surbhi Palande <csurbhi@gmail.com>