tm writes for 7TB disk, cleanup code.
The previous commit was wrong because each tm entry is not 64 bytes but
64 bits. Thus every TM block has 512 TM entries. Every zone of data can
be referenced by 128 blocks of TM entries. Every 1TB (ie 4096 zones) of
data can be referenced by 8 zones of TM entries. 7TB can be addressed by
56 zones. Our HA drive has 64 CMR zones. We can thus completely fit the
TM entries in them! Thus we test with the entire 7TB.

Our assumption for this code is that the extent tree, reverse map tree
stays entirely in memory. Going forward we will have to do something
about the reverse map for GC. The extent tree could be constructed from
the translation table, but there is no way for the same to happen for
the revmap as we do not store any revmap on the disk!!
Currently, for research purpose this will work. But we have to fix this
for ZNS SSDs.

This patch is melded with some cleanup code. The RB tree moves nodes
around as nodes are deleted. Thus the tradition left and right
navigation to remove nodes does not work. So changed the freeing of
translation pages, SIT pages and gc zone and cost tree accordingly.

Signed-off-by: Surbhi Palande <csurbhi@gmail.com>