Fixes for GC.
Testing for multiple GC. Using fstest to do this.
GC was failing because of the following reason:
a) add_revmap_entries calls flush_revmap_block_disk()
It calls a non waiting flush. When this flush completes, it spawns
process_tm_entries() that calls add_translation_entries() and eventuallu
sit_entries_decr()/incr(). On the decr call, zones get reassigned to the
zone tree. So, despite removing a zone from the GC tree at the end of
the GC, the zone gets reassigned this way because of the asynchronous
nature of the call. Another part of the reason was also that
flush_revmap_entries() is called by add_revmap_entries() only on full
pages. So pages that are partially full have entries of zones that are
GCed. We dont want addition of these zones to the trees as they can be
selected wrongly for GC again. So, we make the flushing of revmap and
partial revmap page flush to disk a synchronous process through GC
context.

Signed-off-by: Surbhi Palande <csurbhi@gmail.com>