Debugging stalls seen by threads during GC.
Stalls are seen during GC and concurrent application writes.
Rewritten the gc write phase by giving away the lock immediately when
the gc extents on the list need alteration. Once we know the extent is
perfect, we hold the lock and write it to the disk and update its
metadata all while holding the lock.
However, despite this change, seeing a stall in thread when
flush_write_queue(ctx->write_queues) is called. The dump_stack() shows
that the sub_write_done() takes more than 120 seconds to complete.
In the absence of the GC thread this does not happen!!

Signed-off-by: Surbhi Palande <csurbhi@gmail.com>