TM entries may span more than a single TM page.
When we add tm entries from revmap entries, the tm entry index can
begin anywhere within the TM page. Furthermore, the len can go beyond
the current page. Thus, we need to check for this condition and add a
new tm page to the keystore whenever necessary.

Added a bunch of BUG traps to check for pba, lba and len in the revmap
path and the tm entry addition path!

Also, waiting for the tm entries to be completely written in memory.
Using this in the dtr() to wait for tm pages to be in memory so that
they can be flushed and subsequently freed!

Signed-off-by: Surbhi Palande <csurbhi@gmail.com>