Reads smaller than 4096 block size works.
When smaller reads are issued we create reads for full block sized and
the remaining size we read separately in a new bio and new page. Then we
copy from this page to the actual bio page. We do this when the disk
cannot accept i/o smaller than 4096 bytes ie when the hardware sector
size is 4096 bytes.

Signed-off-by: Surbhi Palande <csurbhi@gmail.com>