Fixed few bugs in GC.
if an extent was split because of lack of space in the gc frontier, then
that extent was not scanned because next was already prepopulated. So
reassigned next to point to the newly added next element. Doing this,
prevents this newly created next from being ignored while writing.

There is a race between lsdm_gc and lsdm_write. so commented out
complete_revmap_flush for now.
There also exists a memory leak (refcount decreases to 0) in the read
code. Added a "read already" flag. But couldnt find the bug yet. This
bug finding is a work in progress.

Signed-off-by: Surbhi Palande <csurbhi@gmail.com>