Keep the tm pages if added at init time.
This way we do not have to read the page from the disk.
If the page is not in memory, its because there are no
valid entries on the disk. Hence such a page can simply be allocated.
This removes unnecessary read of the TM page from the disk during
write end path.

Signed-off-by: Surbhi Palande <csurbhi@gmail.com>