GC enhancement.
We have added another GC mode called concurrent GC.
This is a paused mode, where application I/O get full bandwidth in the
pauses.
We have three watermarks essentially.
A high watermark -> when free zones falls below high watermark, GC needs
to be tiggered, but as long as the middle watermark is not met, GC can
run in a paused mode. You create extents that require reading, pause,
read extents, pause, write extents, free a zone, pause and repeat while
the number of zones is less than high watermark and middle watermark.
You always check after a pause if the number of free zones is still
above middle watermark, otherwise pausing can be dangerous -
applications will stuck for GC and GC pauses for the applications!!
Latency will degrade further in this case. To avoid this, we check the
free zones situtation immediately before and after the pause and verify if we can
still afford to pause

A middle watermark -> when the number of free zones are between the
lower watermark and middle watermark, FG GC needs to continue without
pause. Application I/O need not be blocked until the number of
freezones is above the lower watermark.  After cleaning every zone we call
schedule() so that application thread gets a chance to execute their I/O.
We also wake up the blocked threads after cleaning of 1 zone.

A lower watermark -> when the number of zones falls below this
watermark, the application I/O has to wait in a queue for the free zones
to rise up. The applications I/O is woken up immediately after one zone
is cleaned.

Signed-off-by: Surbhi Palande <csurbhi@gmail.com>