Write reverse translation map based on PBA.
This reduces the random writes that are caused by maintaining a LBA
based translation table.

This also eliminates the need to write reverse map as that is what we
are essentially writing now. Except now we are writing it only every 10
seconds. This was anyway the worst case before as well.

mkfs.ext4, file copy, reads, writes etc work.

Now about to conduct fio tests.

Signed-off-by: Surbhi Palande <csurbhi@gmail.com>