Maximum io len is set to default.
Originally this len was set to
target->max_io_len = BIO_MAX_PAGES >> 1;
This length is not in pages but recorded as sectors.
Linux allows maximum 1MB. There is no reason for us to restrict our I/O
size. This one line reduces the performance of larger I/Os that 128KB.
Remove this.

Signed-off-by: Surbhi Palande <csurbhi@gmail.com>