Write alll cleaned data in the least number of bios.
Linux allows maximum write size of 1MB as the number of pages that can
be packed in a bio is 256. When we clean data, we pack all the valid
blocks and write them at sequential addresses in a zone.
While writing we have to make sure that there is enough space in the
current GC write frontier. This patch collects 1MB worth cleaned data or
the amount of space available in the frontier and then writes the data
out - 1MB at a time. This should help the cleaning of sparsed data ie of
smaller size - such as 4K random writes.

Signed-off-by: Surbhi Palande <csurbhi@gmail.com>