Work in progress.
checking in the code, so that I dont loose it.
This code is work in progress.
Written the algoritm to clean the TM zones.
To do:
a) select the heuristics (greedy for now)
b) insert the trigger - when will this code be called?
Decide - what should you clean first - data or TT?
Cleaning TT has advantage - that you delay data GC and thus allow
more overwriting of the same data before GCing.

Remaining in this algorithm:
a) Update the zone mapping - when the TT blocks are written out to the
new pzone - update the logical zone to physical zone mapping to reflect
this new physical zone.

Signed-off-by: Surbhi Palande <csurbhi@gmail.com>