Combined patches.
1) 1MB writes - in lsdm to compare with on-device STL (DM-Hybrid) should
   write 28.5 GB and then the FG GC should start. To do this, we need
28.5 * 4 = 114 zones free + 6 zones - ie the middle watermark when the
   FG GC will start.
(The 4KB calculation remains).
To achieve this - we removed the command line argument collection. We
hardcoded this in the populate_format.c so that this is the default and
we don't make a mistake.
2) lsdm.c -> diff should be 0 when constructing small bios. This is
   because say the LBA is 2 and the len is 2 sectors for small read, the
PBA is equivalent to the LBA - so the diff does not need to adjust this
from the LBA 0 as we dont read from the PBA corresponding to LBA0 but
from LBA 2.
3) test.sh updated. Wrote a new get_iostat.sh that will get the iostat
   when zfs tests run - or filebench will run.

Signed-off-by: Surbhi Palande <csurbhi@gmail.com>