Read the zone in 1MB chunks during GC.
The maximum bio pages that can be read are 256 ie 1MB.
Read the bio in chunks of 1MB.
This has a constant cost - irrespective of zone utilization.
You can optimize this - by not reading beyond the last valid PBA.
This is our TODO for now.

The next thing on the list is - write the pages in fewest number of
bios. This should considerably improve GC performance - especially for
4KB.

This is an untested patch - i am measuring stuff for my paper.
Will do this after that.

Signed-off-by: Surbhi Palande <csurbhi@gmail.com>